requirements.txt
voila
fastcore
fastai>=0.0.16
pillow>=7.1.0
packaging
ipywidgets==7.5.1
